1. Added support to TUM-VI datasets; 2. Set the tensorflow version to 1.15.5; 3. Added more functions to evaluate_ate_scale.py; 4. Found better parameters for HFNet; 5. Found a critical bug in BaseModel.h;

+++ include/Extractors/HFNetTFModel.h
@@ -32,7 +32,7 @@ public:
     }
 
     bool Detect(const cv::Mat &image, std::vector<cv::KeyPoint> &vKeypoints, cv::Mat &localDescriptors, cv::Mat &globalDescriptors,
-                int nKeypointsNum, int threshold, int nRadius) override;
+                int nKeypointsNum, float threshold, int nRadius) override;
 
     bool IsValid(void) override { return mbVaild; }
